fix-删除后台传文件功能
