+++ tencloud/src/main/resources/application.yml
@@ -2,10 +2,6 @@ server:
   port: 7100
 
 spring:
-  servlet:
-    multipart:
-      max-file-size: 100MB
-      max-request-size: 100MB
   datasource:
     driverClassName: com.mysql.jdbc.Driver
     url: jdbc:mysql://localhost:3306/tenCOSFile?serverTimezone=Asia/Shanghai&characterEncoding=utf8&useSSL=false
