--- tencloud/src/main/resources/templates/index.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<head>
-    <meta charset="UTF-8">
-    <title>Title</title>
-    <link rel="stylesheet" href="https://cdn.staticfile.org/twitter-bootstrap/3.3.7/css/bootstrap.min.css">
-    <script src="https://cdn.staticfile.org/jquery/2.1.1/jquery.min.js"></script>
-    <script src="https://cdn.staticfile.org/twitter-bootstrap/3.3.7/js/bootstrap.min.js"></script>
-    <script>var host = "http://localhost:7100"</script>
-</head>
-<body>
-<script>
-    /**
-     * 上传图片
-     */
-    function uploadImg() {
-
-        /**
-         * formData在jquey中使用需要设置：
-         * processData: false, // 告诉jQuery不要去处理发送的数据
-         * contentType: false // 告诉jQuery不要去设置Content-Type请求头
-         * @type {null}
-         */
-        var fd = new FormData();
-
-        // 第一个参数为controller 接收的参数名称 , input的id
-        fd.append("file", document.getElementById("inputId").files[0]);
-        $.ajax({
-            url: host + "/file/upload",
-            type: "post",
-            data: fd,
-            processData: false,
-            contentType: false,
-            success: function (res) {
-                console.log(res);
-                alert(res);
-                if (res.status.code == 0) {
-                    if (!$('#img').empty()) {
-                        $('#img').empty();
-                    }
-                    // 这一串代码复制不上来 ,截图在下面
-                    $('#img').append(" ![](+res.result[0]+)");
-                } else {
-                    alert("上传失败");
-                }
-            },
-            dataType: "json"
-        })
-    };
-
-    function queryList() {
-        window.location.href = host + '/list?page=0&size=20';
-    }
-</script>
-<div id="img" style="margin: 200px">
-    <input type="button" onclick="queryList()" value="文件列表">
-    <br/><br/>
-    <input type="file" name="text" id="inputId">
-    <br/><br/>
-    <input type="submit" onclick="uploadImg()">
-</div>
-</body>
-</html>
