+++ tencloud/src/main/resources/templates/list.html
@@ -5,7 +5,9 @@
     <link rel="stylesheet" href="https://cdn.staticfile.org/twitter-bootstrap/3.3.7/css/bootstrap.min.css">
     <script src="https://cdn.staticfile.org/jquery/2.1.1/jquery.min.js"></script>
     <script src="https://cdn.staticfile.org/twitter-bootstrap/3.3.7/js/bootstrap.min.js"></script>
-    <script>var host = "http://localhost:7100"</script>
+    <script>
+        var HOST = "http://localhost:7100";
+    </script>
 </head>
 <body>
 <input type="button" onclick="queryList()" value="上传文件">
@@ -20,7 +22,7 @@
 </table>
 <script>
     function queryList() {
-        window.location.href = host;
+        window.location.href = HOST;
     }
 </script>
 </body>
